Fix a bug and compilation warnings
There was a bug where the mesh of the wrong mapchunk
was updated if we placed a node into a different
mapchunk that we were pointing at.
I haven't verified but I belive this bug is fixed now.
The compiler gave us a warning that pos was unused,
so in this case, fixing the warning also fixed the bug.
Great to have a compiler that gives you warnings like these :).