Let mapgen spawn trees for us
We abandon the old generated/not generated
distinction and make it finer, by having
four phases of generation: A mapchunk is
in the "zeroth" phase if no data about
it are present. It is in the first phase
if we have elevation and other info.
A chunk is in the second phase if
all chunks around it are
at least in phase one and it itself
has spawned trees. The third "done"
phase if all chunks around the
given chunk are at least in phase two.

Non-mapgen code doesn't get to see any
phases outside of the done phase.

The second phase ensures that we only place
trees into already generated landscapes
(notice the unwrap call in the spawn_tree function),
and the third phase ensures that mapgen won't
spawn any trees into the mapchunk any more,
as all trees of all chunks around that chunk
have already been generated.
This is needed so that we can generate a mesh
for the chunk and don't have to update it
when other chunks are spawning trees in there.
It also looks weird when suddenly some
tree appears out of nowhere. I didn't like
this property about Minetest's mapgen tbh.

This commit makes mapgen a bit slower, most likely
due to the 5x5x5 + 3x3x3 checks conducted
for every chunk that is being generated.

The introduced stutters might be due to
us generating bigger areas at once.