Add a "Main menu" to the game
Right now its only purpose is
to allow you to exit mouse grab
while the game is running.

Also, I know, the text is aligned
weirdly. Ultimately, we'll need
some functionality to obtain the
width/height of a section and
to use that to position
the section of text,
as well as to obtain the size of
the rectangle in the background
of the text.