Create a mehlon-meshgen crate to make meshgen fast in non-release
Previously, generating a mesh could take up to a few hundred ms per
chunk. This meant that in debug mode you had to wait a couple of seconds
for the meshes to generate at the start and were in general always
waiting for the mesh generation thread to generate the meshes.

Therefore, factor out the mesh generation into a separate crate
and set it's opt-level = 3.

The mesh generation performance is now very similar to release mode.