First commit with network support
We now send/receive packets via TCP,
using bincode for serialization.

Note that you have to fly a little
in order to see any chunks as the
server only sends newly generated
chunks to the clients and the
client usually connects too late.

This commit also uses the second
crate that I've created just for
mehlon:

https://github.com/est31/serde-big-array