Send already generated chunks to new players
Previously, if a player newly joined a server,
the server wouldn't send the player any chunks
that had already been generated.
These chunks would show up as never-generating
holes that the player can fly into but where
nothing is being loaded.
Obviously this is a bug and this commit fixed it.

Optimally, we'd have a faster algorithm,
but for now this algorithm is enough :).