Implement our own (slow) voxel walking algorithm
For some reason, the voxel walking library
didn't produce correct results.
This was quite annoying.
Setting VoxelOrigin::Center vs VoxelOrigin::Corner
makes no real difference: both settings
create the same kind of bug, just for different
positions.
None of the two settings produced bug-free
results for both of the positions:
  * (113.6748, -35.0235, 25.0)
  * (256.52, 13.52, 58.0)

The algorithm we implement is bug-free
in any case.

The next_for_dim function is a bit slow atm
but the focus of this commit is correctness.
The function can be sped up easily.