Use the quinn crate for QUIC based communication
This commit adds a quic_net module that implements
the generic network backend traits for the QUIC
network protocol by using version 0.2 of the
quinn crate.

The quinn crate provides an asynchronous API
based on futures. We do have an async networking
API but it is not based on the futures crate.
Therefore, we put networking into its own
futures-managed thread.
This path was seen as simpler as using
quinn-proto directly.

The current quic implementation still has areas
left to work at:

* supporting graceful connection closing
* removing the panics and bad error messages
* using the length_delimited protocol from tokio::codec