Reduce the number of sqlite transactions
This increases mapgen performance and performance
of anything that writes into the map storage
as each transaction requires a certain overhead.

In measurement data, prior to this change,
storing mapchunks took 27 ms each in the
SqliteStorageBackend::store_chunk function.
With shared transactions, the same function
only takes 50-70 us when no transaction is being
committed/created, and the same approx 27 ms
when a transaction is being committed/created.

Thus, we acheived a big speedup.

There are two bugs in the map storage left
that I found:
 * Some strange off by one error if you remove chunks,
   visible after a restart
 * The sqlite connection isn't closed cleanly.
   This commit makes the situation worse by almost
   always leaving an unfinished transaction behind,
   in the old state, the journal file was only present
   when mapgen was busy for some reason.