Commit transactions on a per-tick basis
This alleviates the bug where the final transaction would
never be commited, basically keeping the sqlite journal
file around all the time and possibly deleting any changes
that happened in that incomplete transaction.

Now, we commit all unfinished transactions each tick
if there are any. For the case that one tick contains
many writes, we still keep a counter around.