Store the mapgen seed in the kv storage of the map
This way, the map will continue to generate
consistent to the saved chunks.

Also, store the mapgen's name. We call our current mapgen "mgv1".
Currently the name does nothing but maybe one day there are
multiple mapgens and the name helps in discerning the
mapgens. We store it from the beginning so that in the
future, nobody has to hardcode "if no mapgen_name is present,
assume mgv1" like stuff.