Add per-player key-value storage
This is useful for storing things like inventory, last position,
skin choice, etc.
ID sources are for future compatibility with multi-id source
systems. E.g. you could imagine having one id source being
singleplayer, another being localauth where the login is managed
directly via the server, a system where the login is managed
across a group of servers, and one where the login is managed
through a centralized service available to all of mehlon.

This is the beginning of a flexible system that hopefully also
will support nick renames and similar, while the question on
uniqueness constraints is left open.