Store the player's position in the map
And use it on load or when a player joins.

For now, the code is only meant to support
singleplayer, as currently the storage location
for the last position of all players is the same.

With proper ID management, and ability to query
player kv entries at runtime, this will change.