Fix bug in server FPS limiting
Previously, if the server was standalone,
a game loop would be so fast that float_delta
would become 0.0. Due to this, fps_cur_term would
always remain at 0.0, and FPS limiting would never
kick in.

This bug has also caused a heavy memory leak
effect where the tick messages in an overwhelmingly
large frequency would be sent to the mapgen thread,
all piling up in the unbounded channel's queue.