Fix a compiler error that for a time was accepted by rustc
There was a compiler bug [1] which made it accept code
with bad delimiters. I wrote commit [2] with a compiler
that included the bug and thus didn't notice the wrongly
placed delimiters. Therefore, on any non-buggy compilers,
the build is broken between that commit and this one.

This commit fixes this.

[1]: https://github.com/rust-lang/rust/issues/58886
[2]: a9b4a2c7fde3d8920519f2152b17c2aae1a55342 "id_src actually isn't needed"