Make merge_or_swap not delete items
Previously, when you clicked on the same item twice in inventory,
it would most often delete the whole stack or parts of it.
This was because it was trying to put the stack into itself and
would override it with the residue. The issue would only not
occur with full stacks as here a swap would've been performed.