Fix bug when you /gime right at login
If you /gime right after login, the inventory
in the server was not the one loaded from disk,
but a newly initialized one. So you'd get your
inventory completely overriden if you did /gime
before the client sent any inventory to the
server, setting the server's copy of the
inventory to a valid state.