Texture-ify the renderer
Change the renderer to handle textures instead of
colors. There is no visual change as each single
texture for now only contains one color. But thanks
to this commit, this is easy to change now in a single
place.

I'm not very happy by the multiple Option<T>'s in the
client that all "only" depend on the game params.
Fixing this is left to future refactors.