Replace MetroHasher with XxHash64
This commit has a non-zero impact as it breaks all maps.
However, I think other commits since the 0.3 already messed
with the mapgen's elevation info.

We mostly do this to get rid of the fasthash dependency.
It is broken on some windows targets [1], and also
requires rand 0.4. This commit rids us of these problems
in place of the twox-hash crate which seems to have better
maintenance.

[1]: https://github.com/flier/rust-fasthash/issues/11