Draw transparent meshes after intransparent ones
Order is important when drawing transparent meshes.

Even oder independent transparency based rendering techniques
assume, as weirdly as it sounds, that you render intransparent
objects first.

This commit reduces the impact of a rendering bug which led to
some chunks not being rendered if seen through a transparent
object that was far enough away. Because in transparent objects
far away, even if we currently only support binary transparency,
either fully transparent or fully intransparent pixels, the
renderer would merge transparent and intransparent pixels to
create half transparent ones, where our discard shader trick
won't work. We could have adjusted the limit of the trick
but we chose this path instead as OIT rendering techniques
already assume a prerendered scene of transparent objects.

The commit only reduces the impact as the scene of intransparent
objects is guaranteed to be drawn before the transparent ones,
however any transparent object as seen through a half-transparent
object (which as outlined above in the current rendering engine
only appear when you are far away) still might or might not
appear, depending on the rendering order.
Ultimately, the scope of the bug is pretty small now as for now
we intend to use transparent objects for only a small subset of
the scene so the *appearance* of objects from far away would not
be affected.