Migrate mehlon-server to edition 2018
I don't like edition 2018 at all but
usage of newer versions of the quinn
library basically requires use of
async keywords, as newer versions
of tokio, futures etc are built with
async in mind. E.g. the read_exact
combinator now takes a reference
instead of an owned structure,
basically requiring you to use the
async keyword.

So I either have to create a new edition
2018 crate that only contains the
networking part, or I switch
mehlon-server to edition 2018.
Again, I don't like edition 2018 but
this is where things are headed to I
guess...

The other crates will remain on
edition 2015 as there is no good
reason to switch to edition 2018
for them.