More proper dig cooldown calculation
With this cooldown calculation, it will reset
if you select a different position during the
digging. This is a preparation for block custom
dig cooldown lengths.

Not sure why, but I notice a slight slowdown
when you hold your LMB to dig stuff continuously.
Maybe it's because it takes an extra frame now?