Rename group:default to group:any
This is a bit undocumented but the code here serves a specific purpose
beyond specifying the default hand tool groups: it also populates
the name id map for the groups, most importantly the id with number 0.

In the client, we use the any group already for this purpose, so
the rename only makes sense.