Store the inventory inside a local variable instead of borrowing
If we move stuff between chests and the inventory,
we will call the set function for the map chunk metadata
handle while the borrow is active. But the metadata
change triggers the on change function of the map,
which sends the metadata change to all players.

This change fixes an already borrowed crash when
moving stuff between chests and the player inventory.

Thus, the new inventory movement logic is now
usable as well as used!