Move most of the client into a mimas-client crate
The idea of this project that this commit starts
is to enable independent server and client crates.

However, for singleplayer, client and server need
to be integrated. The goal is to have this
integration happen in the mimas crate, while the
client crate is generic on the server.

This also neccessitates splitting up of the server
crate to a mimas-common and mimas-server crate,
where everything the client depends on is moved
into the common crate.