Update the nick if it has been changed
The authoritative source for mapping a player's nick
to the id pair is and always will be the auth db.
This db might not be available to people who get
a copy of the map though, e.g. for security reasons.

For them, the map provides a cached version of the nick.
This commit ensures the cached version is updated
and follows nick changes. Currently, there is no
support in the server for changing nicks, but you can
already now edit the db manually.